requirements.txt
fastapi==0.115.0
uvicorn[stnadard]==0.30.6
sqlalchemy==2.0.36
psycopg[binary]=3.2.3
pydantic==2.9.2
python-dotenv==1.0.1
websockets=12.0
anyio=4.4.0